Add diagnostic scripts for Dataset777_GCEF label verification and follow-up analysis
- Created `debug_labels_777.py` to check the integrity of NIfTI label and image files, source TIFF annotations, and preprocessed data.
- Implemented detailed checks for file existence, shape consistency, and unique value counts.
- Added `debug_labels_777_followup.py` for deeper statistical analysis of images, labels, and preprocessed data, including mean, standard deviation, and unique value distributions.
- Introduced checks for configuration files (`plans.json`, `dataset.json`) to ensure proper dataset setup.
- Added a new `setup_prompt.md` file for future setup instructions.

+++ preprocessing_nnUNet_predict_tif.py
@@ -25,7 +25,7 @@ def convert_hdr_to_nii_normalize(input_dir: str,norm_type:str) -> None:
     for hdr_file in tqdm(hdr_files, desc="Process Image to .nii.gz"):
         # Load the file
         img = nib.load(hdr_file)
-        img_arr = img.get_fdata().astype(np.uint8)[:, :, :, 0]
+        img_arr = img.get_fdata()[:, :, :, 0]
 
         # Normalize the image
         img_arr = img_normalize(img_arr, norm_type)
